+++ app/src/main/java/com/github/barriosnahuel/vossosunboton/data/local/Storage.java
@@ -2,7 +2,7 @@ package com.github.barriosnahuel.vossosunboton.data.local;
 
 import android.content.Context;
 import android.content.SharedPreferences;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import java.util.HashSet;
 import java.util.Set;