+++ app/src/main/java/com/github/barriosnahuel/vossosunboton/data/model/Sound.java
@@ -1,8 +1,8 @@
 package com.github.barriosnahuel.vossosunboton.data.model;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.annotation.RawRes;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.RawRes;
 
 /**
  * @author Nahuel Barrios, on 9/4/16.