+++ app/src/main/java/com/github/barriosnahuel/vossosunboton/ui/AbstractActivity.java
@@ -1,8 +1,8 @@
 package com.github.barriosnahuel.vossosunboton.ui;
 
-import android.support.annotation.LayoutRes;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import androidx.annotation.LayoutRes;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.view.LayoutInflater;
 import android.widget.FrameLayout;
 import com.github.barriosnahuel.vossosunboton.R;