+++ app/src/main/java/com/github/barriosnahuel/vossosunboton/ui/DeepLinks.java
@@ -3,8 +3,8 @@ package com.github.barriosnahuel.vossosunboton.ui;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
-import android.support.annotation.NonNull;
-import android.support.annotation.StringRes;
+import androidx.annotation.NonNull;
+import androidx.annotation.StringRes;
 import com.github.barriosnahuel.vossosunboton.R;
 import com.github.barriosnahuel.vossosunboton.util.intent.SafeIntent;