+++ app/src/main/java/com/github/barriosnahuel/vossosunboton/util/file/FileUtils.java
@@ -1,6 +1,6 @@
 package com.github.barriosnahuel.vossosunboton.util.file;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;