+++ app/src/main/java/com/github/barriosnahuel/vossosunboton/util/ui/Feedback.java
@@ -1,8 +1,8 @@
 package com.github.barriosnahuel.vossosunboton.util.ui;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.StringRes;
+import androidx.annotation.NonNull;
+import androidx.annotation.StringRes;
 import android.widget.Toast;
 
 /**